Pull "finished" things into main (#21)
* added unittests for misc and get image

* Add "post" method for RestApiV1 library

* added init with get image and get countries

* Added hashbang line to script

* Added "post" to RestApiV1

* added more GET unit tests

* Adding capability to get images

_get_binary added to RestApiV1

* more unit tests under get_tests

* Updated RestApiV1 "post"

* change line endings to unix-style

* Update .gitignore

updated .gitignore to not include .swp, .pyc

* more unit tests/edits

* Change default REST API url

* Converted RestApiV1 unit tests to use new functions instead of _get

* get roles / get users changed

* Preparing for release 1.1

Release 1.1 allows using External ID and allows attaching subcomponents

* Fixed bug with institution list

* fixed all of misc, some other tests still broken

* Saving state before refactoring _Docket.py

* get components fixes

* checking in before major changes

* post_component

* added patch bulk update

* fixed bulk add

* Updating "Docket" and "Sheet" objects

* Trying to fix line endings to unix instead of dos

* post_test, patch_enable_item, added

* Minor test changes

* patch_enables fixes

* Create test_checklist.txt

* changed test structure

* Update .gitignore

* more stuff

* added unit tests related to tests

* patch subcomponents unit test added

* fixed get_manufacturers

* Long overdue check-in

* added post_subcomponent, removed extra post comp

* added comments/ come tuning

* added comments to all unit tests

* Changes to Encoder and Sheet

* Possibly just line-ending glitches

I'm still trying to get Github Desktop to ONLY use unix line endings, even though it's running on Windows.

* Updating .gitignore

I hope I did this right!

* Updated Unit Tests for RestApiV1

* Fixed several tests that were skipped because they weren't working
* Created new LoggedTestCase class that performs some logging on setUp/tearDown automatically, and changed existing unit tests to use it
* Moved unit tests to a UnitTest folder on the same level as SystemTest
* Added docstrings to test cases

* Changes to RestApiV1 and UnitTests

* Added a lot of docstrings
* Changed the name of some RestApi functions for consistency
* Added image tests
* RestApi "request" now raises "insufficient permissions" when applicable
* Probably a lot of things I can't think of right now

* Unit test fixes

Fixed image post unit test
Updated copyright year on all files that had one

* Update UnitTests

Updated "base" UnitTest.testcase to automatically log erros, so try/except blocks are no longer needed in test cases to write this info into the logs.

* Continuing development on version 1.2

* Renamed a directory, so it looks bigger than it really is.

* Modules for writing Excel sheets and for managing HWItems through a consistent interface

I had the Excel sheet writer mostly working in a "scratch" folder, but I finally put it in the library for general use. I added a bit of code to use the ImageFont module out of PIL (Python Image Library) to get a better estimate of the width of columns in Excel sheets, instead of merely counting characters. It appears to be working really well; the sheets come out looking reallly nice.

The new HWItem module is there to consolidate a lot of various bits of code that do things with Items but not in a really consistent manner. It serves as a go-between between the format things come out of the HWDB REST API as, and the format the Encoder puts things in after it reads a sheet.

* Rewriting Docket

* pre-release for v1.2

* Unit Tests: added tests for 'enabled' bugs

* Fixing bugs for 1.2

* Bug fixes

* More bug fixes plus one example

* Cleaning up display and fixing bugs

* Changed how jobs get executed

* Updated demo

* Cleaned up display when executing jobs

* Updated version number

* Added 'location' endpoints to REST API interface

* Handle change in API from "enabled" to "status"

* reduced colors in images

changed image-to-text to use only 256 colors instead of 24-bit, since Mac terminals don't seem to support 24-bit. Should try to detect this in the future instead.

* Update Configuration to work with OpenSSL 3

* Removed old code

* Fixed subcomponents, added item images

* Added tutorial examples

* Fixed ability to reference <null> subcomponent on new items

* Update tutorial examples

* Fixed accidental swap of Lesson 4 examples

* detailed unittest responses in terminal

* verbose messaging for unit tests

* verbose unit test additions

added summary at the end, added base_test_case.py to do the set up and tear down for most unit tests

* test numbers added

* Deleted unnecessary output files from unit tests

* Fixing Urbas's merge

I'm still not sure this will be right.

I think what happened was that he "merged" in such a way that he kept his versions of files instead of taking mine, and now it never seems to see my branch and his branch as different, even though THEY ARE!! I manually deleted his /lib and copied mine, but it doesn't show up in changes. So I don't know what is going to happen now.

* added HWDBUtility system test that tests having subcomponents of the same type as the component

* Added locations and multiple images

* fixed "arrival"->"arrived" bug

* app now outputs a summary sheet of items uploaded

* removed extra print statements

* fixed label generation to better recognize items that aren't in the database yet

* removed code to create labels since it is now in __init__

* Fixed issue with pdf label writer

there was a bug where if the list of part_ids had duplicates, it would get the data multiple times, but not be able to unpack the results for the second one. Changed it so that the incoming list would be filtered to make them all unique.

* changed release number

* Adding ability to upload encoders into the HWDB

* Created workaround for requests seg fault problem with requests 2.32 in python 3.12

* Module now shares a single ThreadPoolExecutor

* fixed locations bug

* added label customization

* added standalone label script

* fixed copying default config files

* changed default custom labels, and added font face and free text to labels

* Changed some label templates

* updated version number

* PDF Labels changed to accommodate serial number

* Made _meta optional in Items

* switched authentication to htgettoken

* enabled switching between certificate and htgettoken authentication

* Shipping GUI

* Updated shipping application

* Fixed error when updating with manufacturer=none

* Changed how token/certificate works

* added --labels option to hwdb-upload

* Show URL if htgettoken takes too much time

* changed initial timeout from 2 seconds to 5 seconds when calling htgettoken

* hacked htgettoken to flush its output while still running, so we can use it to display URL

* Update hwdb-htgettoken-alt

* updated version

* Updated shipping gui application

* Updated Shipping GUI app

* clean up code

* Updated Shipping application

* Fixed Receiving and Transit

* Configuration changes

* Updated RestApi, Configuration, Shipping

* Inserted new page in preshipping

* Renumbered preshipping pages

* removed "close" button in shipping app (temporarily)

* made new part_details widget to show on several pages

* added part details to shipping workflow

* rempved a print statement

* bug fixes in shipping application

* Refactoring

Changed the names of some variables and files

* Rearranged file structure

* More restructuring

* Changed how locations widget gets location data

* Update version number

* Removed duplicate unit tests

* Update select_pid.py

---------

Co-authored-by: ekka0002 <141053176+ekka0002@users.noreply.github.com>